# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bloom Together API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List love notes
      parameters:
        - name: pinned
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Write a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /notes/{id}/pin:
    patch:
      operationId: toggleNotePin
      tags: [notes]
      summary: Toggle pin on a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}/react:
    post:
      operationId: reactToNote
      tags: [notes]
      summary: React to a note with a heart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  # TASKS
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List shared tasks
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: completed
          in: query
          schema:
            type: boolean
        - name: assignedTo
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/complete:
    patch:
      operationId: completeTask
      tags: [tasks]
      summary: Toggle task completion (only assigned user may complete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated task with optional newFlower if a bloom was triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteTaskResult"
...
  /tasks/flowers/{flowerId}/gift:
    post:
      operationId: giftBloomFlower
      tags: [tasks]
      summary: Gift a bloom flower to your partner
      parameters:
        - name: flowerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftFlowerInput"
      responses:
        "200":
          description: Gifted flower
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloomFlower"

  # FOCUS SESSIONS
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        favoriteFlower:
          type: ["string", "null"]
        favoriteActivity:
          type: ["string", "null"]
        createdAt:
          type: string

    # BLOOM SPACE SCHEMAS
...
    NoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        style:
          type: string
          enum: [sticky, letter, floral]
    NoteUpdate:
      type: object
      properties:
        content:
          type: string
        style:
          type: string
          enum: [sticky, letter, floral]
        isPinned:
          type: boolean
    Note:
      type: object
      required: [id, bloomSpaceId, content, createdBy]
      properties:
        id:
          type: integer
        bloomSpaceId:
          type: integer
        content:
          type: string
        style:
          type: string
        isPinned:
          type: boolean
        heartCount:
          type: integer
        createdBy:
          type: integer
        createdAt:
          type: string
        creator:
          $ref: "#/components/schemas/User"

    # TASK SCHEMAS
    TaskInput:
      type: object
      required: [title, assignedTo]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        assignedTo:
          type: integer
        dueDate:
          type: ["string", "null"]
        category:
          type: string
          enum: [personal, study, work, travel, relationship, custom]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        category:
          type: string
          enum: [personal, study, work, travel, relationship, custom]
        completed:
          type: boolean
    Task:
      type: object
      required: [id, bloomSpaceId, title, completed]
      properties:
        id:
          type: integer
        bloomSpaceId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        category:
          type: string
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        assignee:
          $ref: "#/components/schemas/User"
    CompleteTaskResult:
      type: object
      required: [task]
      properties:
        task:
          $ref: "#/components/schemas/Task"
        newFlower:
          $ref: "#/components/schemas/BloomFlower"
        progressToNext:
          type: integer
...
    BloomFlower:
      type: object
      required: [id, bloomSpaceId, earnedBy, flowerType, gifted, createdAt]
      properties:
        id:
          type: integer
        bloomSpaceId:
          type: integer
        earnedBy:
          type: integer
        flowerType:
          type: string
          enum: [rose, peony, tulip, daisy, cherry_blossom]
        gifted:
          type: boolean
        giftedTo:
          type: ["integer", "null"]
        giftMessage:
          type: ["string", "null"]
        giftedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        earner:
          $ref: "#/components/schemas/User"
        recipient:
          $ref: "#/components/schemas/User"
    GiftFlowerInput:
      type: object
      required: []
      properties:
        message:
          type: ["string", "null"]

    # FOCUS SESSION SCHEMAS
...