# lib/api-zod/src/generated/types/noteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bloom Together API
 * OpenAPI spec version: 0.1.0
 */
import type { NoteUpdateStyle } from './noteUpdateStyle';

export interface NoteUpdate {
  content?: string;
  style?: NoteUpdateStyle;
  isPinned?: boolean;
}